Fixed potential (minor) possible null dereferencing problems found by the PVS-Studio static code analyzer.
In most cases, the possible root cause of null pointer dereferencing is event handler invocation. In practice,
it's almost always safe to do a check-and-call sequence but speaking formally we have to make a local copy
of the handler before making a call. The null coalescing operator does it automatically and event invocation
looks much shorter in this case.