fix: do not attempt to exit early during Binding update
Previously, during a Binding update, we attempted to shortcut the process by
comparing values by reference. This could cause problems when the reference
did not change but the its content did, if for instance the Binding uses a
converter which uses the content to create some text.
Now we notify the target property in all cases and give the responsibility to
check if the update is relevant to the DependencyObject class.