fix: update format of the code generated by the XamlPreprocessor
Before this commit, when a xaml file without backend C# code was being
processed by the XamlProcessor, we generated a class in order to fall back
to the case where there was a backing C# file. For the following xaml code :
<ResourceDictionary>
    <Style x:Key="key1" />
</ResourceDictionary>
we ended up with a generated code looking as follows:
public class __Generated_Name_For_The_Xaml_File : ResourceDictionary
{
    public __Generated_Name_For_The_Xaml_File()
    {
        InitializeComponent();
    }

    public void InitializeComponent() { ... }
}
This could become an issue in rare scenario where the root of the xaml file
is a sealed type, like ControlTemplate.
In the new implementation, we do not generate this additional class and
handle the xaml component's creation directly in the factory type generated
alongside the removed class (the factory type was generated before as well,
but instead of taking care of building the component, it was only instantiating
via reflection the generated type).

BREAKING CHANGE: as the generated types were public, it is possible to have
compiling issues if they were explicitely referenced in user code. In such a
case, using the factory type to generate the content is the most appropriate
workaround.