refactor: change the mechanism used to store routed event handlers
Previously, handlers were stored inside the event manager associated to the routed event.
This could cause issues when more than one routed event has to be mapped to the same dom
event. Indeed, in such a case we had two options:
* Use a single event manager for all concerned routed events
  In that case, we were not be able to know to which routed event a handler was associated,
  as handlers from all routed events were merged in a single list.
* Use as many event managers as routed events
  In that case, every time the DOM event is fired, every single event manager is notified
  and all the handlers end up being invoked.
Now, each UIElement owns a store that holds the handlers. This store is independent from the
event manager. This allows to create one event manager per DOM event, and then dispatch the
event to any routed event when needed.