fix: fix Binding with RelativeSource (Mode=TemplatedParent) in style setter
Previously, the compiler was attempting to resolve the source of a Binding with a relative
source using the TemplatedParent mode at compile time. This is not possible when the Binding
is declared in a style's setter as the context provided by the xaml code does not allow to
determine the templated parent. Now this step is skipped when processing a style's setter,
and the binding source has to be resolved later.