fix: exclude implicit values from dependency property storage
Previously, an implicit value was returned when calling DependencyObject.GetValue(...).
However, Silverlight only support implicit values for the FrameworkElement.Style property,
and calls to FrameworkElement.Style always return null (the Style property default value)
when an implicit style is in use.
To enfore this behavior, implicit values are not stored in the dependency property storage
anymore, and FrameworkElement is now responsible for applying and storing implicit styles.