fix: prevent GC from collecting the CanExecuteChanged handler of a button's command (#639)
Previously, when a button was using a command that implements the CanExecuteChanged event with a
weak event pattern (by storing the handler in a weak reference), the handler could end up being collected
too early by the garbage collector. We add a strong reference to the handler in the ButtonBase implementation
to ensure that the handler stays alive as long as the ButtonBase is alive.