Memory consumption improvements (expected: 10-15+%)
Memory consumption improvements (expected: 10-15+%)

1. JavascriptCallback
A. Now, I hold a dictionary with the Delegate being the key.  The advantage here is that for a certain Delegate, only one callback is generated (and subsequently returned) - less leaks
B. Added .Remove(Delegate), so you can manually remove the Delegate from the callback collection once you don't need it anymore
C. You can use RuntimeHelpers.RemoveJavascriptCallback to remove such a callback yourself, from client code
D. Overload for InteropImplementation.GetVariableStringForJS for delegates

2. Error Callback (INTERNAL_InteropImplementation)
A. Setting to turn ErrorCallback ON/OFF (OFF by default)
B. The idea here is that when using ErrorCallback, this will leak strings. So, just use it if/when you get a JS Exception (you'll see that in the console)

3. SynchronizedStore
A. Use a dictionary instead of a list.
B The rationale is -- in the old code, the List would only grow, Clean() would just set the _items[index] to default.