fix: set native focus synchronously
Previously, the focus was being updated after a 1 ms delay, which could lead to
unexpected behavior in certain scenarios. For example, if the 'Control.Focus'
method was called during a 'KeyDown' event after the Tab key was pressed, the
focus would first move to the next focusable element in the DOM due to tab
navigation, and then move to the element that called the 'Control.Focus' method.
This is the opposite of the expected behavior, where the focus should move to the
element that called the 'Control.Focus' method first, and then move to the next
focusable element in the DOM.