fix: update tabindex management
Previously, setting 'IsTabStop' to false was translated to a tabindex of -1 in
html, which would make the element focusable. This resulted in receiving unwanted
'focus' event when clicking on elements. To prevent this from happening, the
tabindex attribute is now removed from html elements when IsTabStop is set to
false.