feat: adjust behavior of ScrollBar.Debounce property
Before this commit, the Debounce property could only be applied directly to
a ScrollBar or globally thanks to the Settings.ScrollDebounce property. This was
very inconvenient because scrollbars are almost always declared in templates
that can't be easily accessed. Now, a scrollbar attempts to get the debounce
interval from its templared parents.
This makes cases like this possible (assuming the ListBox declares a ScrollViewer
in its template):
<ListBox ScrollBar.Debounce="0:0:0.100" />