feat: enhance Storyboard for better WPF compatibility
This commit adds support for multiple targets on Storyboards. It also implements several new
methods, such as GetCurrentIteration(), GetCurrentProgress(), GetIsPaused(), and Remove().
Finally, the Seek() and SeekAlignedToLastTick() behavior has been changed to match WPF instead
of Silverlight; when the seek operation is based on the BeginTime, any begin time will be
automatically skipped.